Replace small integer types (`u8`, `u16`) with `u32` as defensive measure (#1348)
* Replace small integer types (`u8`, `u16`) with `u32` as defensive measure

An 11-year-old bug in Android's ART AOT compiler for ARM32 causes incorrect sign-extension of `Short` (16-bit) JNI return values, which can lead to UniFFI checksum mismatches at runtime (#1339).

While the upstream fix is being pursued separately, this change defensively widens all small integer types (`u8`, `u16`) to `u32` across struct fields, function parameters, return types, and enum reprs — particularly those on or near the UniFFI surface. This eliminates the possibility of hitting the AOT sign-extension bug if any of these types are ever exposed through UniFFI bindings.

Byte data (`Vec<u8>`, `&[u8]`, `[u8; N]`) used for request bodies, cryptographic keys, and certificates is left unchanged. Internal `usize` usage tied to `rusqlite` APIs and array indexing is also unchanged, as these types don't cross the FFI boundary.

* Add changelog entry for small integer type widening (#1339)

* Fix missed `u16`/`UShort` in integration tests and Kotlin wrappers

Update `WpRequestResult.kt` status code fields from `UShort` to `UInt`,
`WpRequestExecutor.kt` response code conversion from `toUShort()` to
`toUInt()`, and `mock.rs` test helper parameter from `u16` to `u32`.

* Fix remaining `u16` in integration tests and unnecessary cast clippy warning

* Fix `UInt16` in Swift wrappers to match widened Rust types

* Fix `UInt16` in Swift test helper to match widened Rust types

* Mark changelog entry as breaking change

* Use `as u64` consistently in `SyncResult::no_op`

* Change test helper parameter from `u16` to `u32` to avoid unnecessary cast

+++ native/swift/Sources/wordpress-api/Extensions.swift
@@ -19,7 +19,7 @@ extension WpNetworkResponse {
 
         self = WpNetworkResponse(
             body: data,
-            statusCode: UInt16(response.statusCode),
+            statusCode: UInt32(response.statusCode),
             responseHeaderMap: try WpNetworkHeaderMap.fromMap(hashMap: response.httpHeaders),
             requestUrl: request.url(),
             requestMethod: request.method(),
